backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dealwise-backend"
version = "0.1.0"
description = "DealWise AI - Meeting Intelligence Platform Backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",

    # AI/LLM
    "anthropic>=0.34.0",
    "openai>=1.40.0",
    "tiktoken>=0.7.0",

    # Transcription
    "deepgram-sdk>=3.5.0",

    # AWS
    "boto3>=1.34.0",
    "aioboto3>=13.0.0",
    "aiobotocore>=2.13.0",

    # Utilities
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "structlog>=24.4.0",
    "python-dateutil>=2.9.0",
    "tenacity>=8.5.0",

    # File processing
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "flower>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "ICN", "PIE", "T20", "SIM", "TCH"]
ignore = ["S101", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
